Move data structures to particular packages
