+++ datastructures/list/list.go
@@ -1,4 +1,4 @@
-package datastructures
+package list
 
 import (
 	"errors"
