+++ datastructures/stack/stack_test.go
@@ -1,4 +1,4 @@
-package datastructures
+package stack
 
 import (
 	"errors"
